requirements.txt
requests
pandas
pyarrow
google-cloud-storage
sqlalchemy
aiohttp
apache-airflow
google-cloud-bigquery
apache-airflow-providers-apache-spark
scikit-learn
google-cloud-bigquery
pandas-gbq